+++ Reinvest/Investments/src/Application/UseCases/StartInvestment.ts
@@ -1,5 +1,7 @@
 import { UUID } from 'HKEKTypes/Generics';
+import { InvestmentCreated, TransactionEvents } from 'Investments/Domain/Transaction/TransactionEvents';
 import { InvestmentsRepository } from 'Investments/Infrastructure/Adapters/Repository/InvestmentsRepository';
+import { storeEventCommand } from 'SimpleAggregator/EventBus/EventBus';
 
 class StartInvestment {
   private readonly investmentsRepository: InvestmentsRepository;
@@ -37,9 +39,31 @@ class StartInvestment {
       return false;
     }
 
-    const isCorrectlyUpdated = await this.investmentsRepository.updateInvestment(investment, approveFees);
+    const investmentData = investment.toObject();
 
-    return isCorrectlyUpdated;
+    const events = [
+      <InvestmentCreated>{
+        id: investmentData.id,
+        kind: TransactionEvents.INVESTMENT_CREATED,
+        date: new Date(),
+        data: {
+          profileId,
+          accountId: investmentData.accountId,
+          portfolioId: investmentData.portfolioId,
+          parentId: investmentData.parentId,
+          amount: investmentData.amount,
+        },
+      },
+      storeEventCommand(profileId, 'InvestmentProcessStarted', {
+        accountId: investmentData.accountId,
+        amount: investmentData.amount,
+        tradeId: investmentData.tradeId,
+        origin: investmentData.recurringInvestmentId ? 'Recurring investment' : 'Direct deposit',
+        investmentId: investmentData.id,
+      }),
+    ];
+
+    return this.investmentsRepository.updateInvestment(investment, approveFees, events);
   }
 }