+++ Reinvest/Notifications/src/Adapter/Database/NotificationsSchema.ts
@@ -1,3 +1,5 @@
+import { JSONObject, UUID } from 'HKEKTypes/Generics';
+
 export interface NotificationsTable {
   accountId: string | null;
   body: string;
@@ -13,4 +15,34 @@ export interface NotificationsTable {
   onObjectType: string | null;
   profileId: string;
   uniqueId: string | null;
+}
+
+export interface StoredEventsTable {
+  dateAccountActivity: Date | null;
+  dateAnalytics: Date | null;
+  dateCreated: Date;
+  dateEmailed: Date | null;
+  dateInApp: Date | null;
+  datePushed: Date | null;
+  id: string;
+  kind: string;
+  payloadJson: JSONObject;
+  profileId: string;
+  status: string;
+}
+
+export interface AccountActivitiesTable {
+  accountId: string | null;
+  activityDate: Date;
+  activityName: string;
+  dataJson: JSONObject | null;
+  hash: string;
+  profileId: string;
+}
+
+export interface RegisteredPushDevicesTable {
+  dateUpdated: Date;
+  deviceId: string;
+  id: UUID;
+  profileId: UUID;
 }