+++ Reinvest/Trading/src/Port/Queue/EventHandler/CreateTradeHandler.ts
@@ -29,6 +29,7 @@ export class CreateTradeHandler implements EventHandler<DomainEvent> {
       subscriptionAgreementId: event.data.subscriptionAgreementId,
       portfolioId: event.data.portfolioId,
       parentId: event.data.parentId,
+      userTradeId: event.data.userTradeId,
     };
     const tradeSummary = await this.createTradeUseCase.createTrade(tradeConfiguration);