+++ Reinvest/Verification/src/Providers/EventBusProvider.ts
@@ -1,12 +1,12 @@
 import { ContainerInterface } from 'Container/Container';
-import { EventBus, SimpleEventBus } from 'SimpleAggregator/EventBus/EventBus';
+import { EventBus, SimpleEventBus, STORE_EVENT_COMMAND } from 'SimpleAggregator/EventBus/EventBus';
 import { SendToQueueEventHandler } from 'SimpleAggregator/EventBus/SendToQueueEventHandler';
 import { Verification } from 'Verification/index';
-import { VerifyAccount } from 'Verification/IntegrationLogic/UseCase/VerifyAccount';
-import { VerifyAccountForInvestmentHandler } from 'Verification/Port/Queue/EventHandler/VerifyAccountForInvestmentHandler';
 import { NotifyAccountNotVerifiedEventHandler } from 'Verification/IntegrationLogic/Events/NotifyAccountNotVerifiedEventHandler';
-import { MarkSensitiveDataUpdatedHandler } from 'Verification/Port/Queue/EventHandler/MarkSensitiveDataUpdatedHandler';
+import { VerifyAccount } from 'Verification/IntegrationLogic/UseCase/VerifyAccount';
 import { VerifierRepository } from 'Verification/IntegrationLogic/Verifier/VerifierRepository';
+import { MarkSensitiveDataUpdatedHandler } from 'Verification/Port/Queue/EventHandler/MarkSensitiveDataUpdatedHandler';
+import { VerifyAccountForInvestmentHandler } from 'Verification/Port/Queue/EventHandler/VerifyAccountForInvestmentHandler';
 
 export default class EventBusProvider {
   private config: Verification.Config;
@@ -28,6 +28,7 @@ export default class EventBusProvider {
       'PrincipalVerificationMadeDecision',
     ]);
     eventBus.subscribeHandlerForKinds(SendToQueueEventHandler.getClassName(), [
+      STORE_EVENT_COMMAND,
       'AccountVerifiedForInvestment',
       'AccountBannedForInvestment',
       'CreateNotification',